Adding latest version of titlesec as file
I just noticed that whenever I compile the paper that the section numbers are missing. I didn't mind at first since I though that this is what the template wants ... until I got review papers.

Looking into it, the cause is a bug in the `titlesec` package that was added as dependency for "compact" title headings. That package had a bug was shipped with the Ubuntu 16.04 which is a Long Term Release.

As I didn't find an easy workaround the easiest fix 
(also for the users) is to simply include the latest fixed 
version 2.10.2 with this template.